Phase 16.x: Production optimizations and bug fixes
Modified Files (10):
1. RealtimeITagControl/CamViewerCore.cs
2. RealtimeITagControl/RealtimeITagControl.cs
3. RealtimeITagControl/Selection/GeometryUtils.cs
4. RealtimeITagControl/Rendering/OpenGLRenderMode.cs
5. RealtimeITagControl/ITagManager.cs
6. RealtimeITagControl/MPF/MPFParser.cs
7. RealtimeITagControl/MemoryOptimizer.cs
8. RealtimeITagControl/Rendering/RenderSettings.cs
9. RealtimeITagControl/TagDefinitions.cs
10. RealtimeITagControl/UI/ProgramInfoPanel.cs

Changes Summary:

CamViewerCore.cs:
1) Line 215: Panel size 800x600 → 968x630 (HMI screen fit)
2) Line 236: DoubleBuffered true → false (OpenGL conflict fix)
3) Line 1400: DrawTextOverlays() commented (duplicate text removal)
4) Line 1741: base.Invalidate() commented (unnecessary redraw)
5) Debug.WriteLine removed (production cleanup)
6) Line 2685: Settings path fixed to C:\HK\CamViewerPOC\ with auto-create

RealtimeITagControl.cs:
7) Line 877: camViewerControl.Invalidate() - Not found (already optimized)
8) Line 942: camViewerControl.Invalidate() - Not found (already optimized)
9) Line 1439,1461: UpdateContourTotalProgress() commented (HKSTO white reset fix)
10) Line 771: ActLineCode TrimEnd('\0','\r','\n',' ','\t') added (Siemens special chars)

GeometryUtils.cs:
11) Debug.WriteLine removed (production cleanup)

OpenGLRenderMode.cs:
12) Default mode: Default(0) → DirtyRegion(1) (flickering prevention)

Rationale:
- #2: GDI+ DoubleBuffered conflicts with OpenGL SwapBuffers on device
- #3: OpenGL text rendering handles this, prevents duplication
- #4: renderPanel.Invalidate() is sufficient, base call unnecessary
- #9: Prevents incorrect white reset for HKSTO end codes (GC11/GC12/GC13)
- #10: Siemens ITag includes null/CR/LF/tab characters in GCode strings
- #12: Full screen refresh causes severe flickering on device with CycleReadTime

Testing Status:
- All changes tested on local device
- Normal operation confirmed
- Flickering eliminated with DirtyRegion default

Build: Success
Changes: 10 files, +100 insertions, -135 deletions (net -35 lines)

+++ Phase15/RealtimeITagControl/ITagManager.cs
@@ -440,10 +440,10 @@ namespace RealtimeITagControl
                             tagData.ActLineNum = Convert.ToInt32(value);
                             break;
                         case TagDefinitions.SEARCH_PART:
-                            tagData.SearchPart = Convert.ToInt32(value);
+                            tagData.SearchPart = Convert.ToDouble(value);
                             break;
                         case TagDefinitions.SEARCH_CONT:
-                            tagData.SearchContour = Convert.ToInt32(value);
+                            tagData.SearchContour = Convert.ToDouble(value);
                             break;
                         case TagDefinitions.DIR_TYPE:
                             tagData.DirType = (TagDefinitions.DirectionType)Convert.ToInt32(value);
